refactor(dispatcher): retire Tier-4 legacy proxy into router (#624)
Absorb dispatcher/proxy.py:resolve_slot into router.py as the module-level
resolve_by_capability(), behavior-preserving, then delete proxy.py. Safe
redo of #649 (closed): proxy.py was NOT dead — four capability endpoints
(/v1/{embeddings,rerankings,audio/speech,images/generations}) plus FLM
name:tag -> npu routing depend on its path/name pins, because the registry
doesn't bind their model ids (kokoro-v1, sdxl-turbo, bge-reranker-*).

- Port rules 1-6 + SLOT_ALIASES operator slot-name addressing verbatim into
  router.resolve_by_capability; relocate LegacyResolutionFailed (its
  dispatch.legacy_unresolved code is preserved — it's an error contract).
- dispatch() Step 4 reframed "legacy fallback" -> "capability/path routing";
  resolution_path "legacy_slot:*", the details.legacy_error key, and the
  LegacyResolutionFailed -> NoRouteFound mapping are all kept unchanged.
- Repoint every test off hal0.dispatcher.proxy onto router; add a rule-5
  (name:tag -> npu) characterization test that was missing from the net.

No behavior change. Dispatcher/routing suite green (223 passed). Full suite:
3318 passed; the 5 failures + 4 errors are pre-existing and environment-
specific (proxmox-guest detection, root-user writability, moonshine audio
dep) — confirmed identical on pristine main.
