+++ src/hal0/dispatcher/__init__.py
@@ -5,10 +5,13 @@ to forward each OpenAI-compatible API request.  It handles:
   - Registry-exact routing (model_id → slot binding)
   - Cold-cache prefetch with single-flight coalescing
   - Remote upstream fallback (OpenRouter, Anthropic, etc.)
+  - Capability/path routing (:func:`hal0.dispatcher.router.resolve_by_capability`),
+    the last-resort step that pins embed/rerank/tts/image by path and FLM
+    ``name:tag`` models to the npu slot
   - Structured decision logging to journald
 
-The legacy proxy shim (proxy.py) is kept during Phase 1 and will be
-absorbed post-v0.2.
+The legacy ``proxy.py`` shim has been absorbed into ``router.py`` — its
+``resolve_slot`` heuristics now live in ``router.resolve_by_capability``.
 
 Port targets: haloai lib/dispatcher.py (617 lines), lib/proxy.py.
 See PLAN.md §3 and §5 Tier 2/3.
